refactor(test): simplify and improve readability of auth endpoint tests
